+++ docs/changelog.md
@@ -22,10 +22,10 @@ title: Changelog
   have been deleted by another client. Instead of crashing with a 500 error, the server returns a 410 Gone response
   with a `StaleComponentState` error code, allowing the client to handle the situation appropriately 
   (e.g., by removing the component).
-  
 
 ### Changed
 - Pickling/Unpickling now checks variables against a "safe list" of types.
+- rename Alpine store `tetra_status` to `tetraStatus`
 
 ### Removed
 - Remove deprecated `@public.subscribe` decorator